Move CHECK_CODES to version-specific submodules
Required moving some code around, and giving each STIX subversion its own
JSONError class.
