+++ stix2validator/util.py
@@ -11,6 +11,8 @@ from appdirs import AppDirs
 import requests_cache
 
 from .output import error, set_level, set_silent
+from .v20.enums import CHECK_CODES as CHECK_CODES20
+from .v21.enums import CHECK_CODES as CHECK_CODES21
 
 DEFAULT_VER = "2.1"
 
@@ -18,6 +20,9 @@ CODES_TABLE = """
 The following is a table of all the recommended "best practice" checks which
 the validator performs, along with the code to use with the --enable or
 --disable options. By default, the validator checks all of them.
+
+This table is for STIX version {}. For older versions, please refer to
+https://stix2-validator.readthedocs.io/en/latest/best-practices.html.
 
 +------+-----------------------------+----------------------------------------+
 | Code | Name                        | Ensures...                             |
@@ -58,22 +63,22 @@ the validator performs, along with the code to use with the --enable or
 |      |                             | attack_resource_level vocabulary       |
 | 213  | identity-class              | certain property values are from the   |
 |      |                             | identity_class vocabulary              |
-| 214  | indicator-label             | certain property values are from the   |
-|      |                             | indicator_label vocabulary             |
+| 214  | indicator-types             | certain property values are from the   |
+|      |                             | indicator_types vocabulary             |
 | 215  | industry-sector             | certain property values are from the   |
 |      |                             | industry_sector vocabulary             |
-| 216  | malware-label               | certain property values are from the   |
-|      |                             | malware_label vocabulary               |
-| 218  | report-label                | certain property values are from the   |
-|      |                             | report_label vocabulary                |
-| 219  | threat-actor-label          | certain property values are from the   |
-|      |                             | threat_actor_label vocabulary          |
+| 216  | malware-types               | certain property values are from the   |
+|      |                             | malware_types vocabulary               |
+| 218  | report-types                | certain property values are from the   |
+|      |                             | report_types vocabulary                |
+| 219  | threat-actor-types          | certain property values are from the   |
+|      |                             | threat_actor_types vocabulary          |
 | 220  | threat-actor-role           | certain property values are from the   |
 |      |                             | threat_actor_role vocabulary           |
 | 221  | threat-actor-sophistication | certain property values are from the   |
 |      |                             | threat_actor_sophistication vocabulary |
-| 222  | tool-label                  | certain property values are from the   |
-|      |                             | tool_label vocabulary                  |
+| 222  | tool-types                  | certain property values are from the   |
+|      |                             | tool_types vocabulary                  |
 | 241  | hash-algo                   | certain property values are from the   |
 |      |                             | hash-algo vocabulary                   |
 | 242  | encryption-algo             | certain property values are from the   |
@@ -102,7 +107,7 @@ the validator performs, along with the code to use with the --enable or
 | 302  | extref-hashes               | external references SHOULD have hashes |
 |      |                             | if they have a url                     |
 +------+-----------------------------+----------------------------------------+
-"""
+""".format(DEFAULT_VER)
 
 
 class NewlinesHelpFormatter(RawDescriptionHelpFormatter):
@@ -162,7 +167,9 @@ def parse_args(cmd_args, is_script=False):
     parser.add_argument(
         "--version",
         dest="version",
-        help="The version of the STIX specification to validate against."
+        default=DEFAULT_VER,
+        help="The version of the STIX specification to validate against (e.g. "
+             "\"2.0\")."
     )
 
     # Output options
@@ -352,59 +359,21 @@ class ValidationOptions(object):
         set_level(self.verbose)
         set_silent(self.silent)
 
+        if self.version == '2.0':
+            check_codes = CHECK_CODES20
+        else:  # Default version
+            check_codes = CHECK_CODES21
+
         # Convert string of comma-separated checks to a list,
         # and convert check code numbers to names
         if self.disabled:
-            self.disabled = self.disabled.split(",")
-            self.disabled = [CHECK_CODES[x] if x in CHECK_CODES else x
+            self.disabled = self.disabled.split(',')
+            self.disabled = [check_codes[x] if x in check_codes else x
                              for x in self.disabled]
         if self.enabled:
-            self.enabled = self.enabled.split(",")
-            self.enabled = [CHECK_CODES[x] if x in CHECK_CODES else x
+            self.enabled = self.enabled.split(',')
+            self.enabled = [check_codes[x] if x in check_codes else x
                             for x in self.enabled]
-
-
-# Mapping of check code numbers to names
-CHECK_CODES = {
-    '1': 'format-checks',
-    '101': 'custom-prefix',
-    '102': 'custom-prefix-lax',
-    '111': 'open-vocab-format',
-    '121': 'kill-chain-names',
-    '141': 'observable-object-keys',
-    '142': 'observable-dictionary-keys',
-    '149': 'windows-process-priority-format',
-    '150': 'hash-length',
-    '2': 'approved-values',
-    '201': 'marking-definition-type',
-    '202': 'relationship-types',
-    '203': 'duplicate-ids',
-    '210': 'all-vocabs',
-    '211': 'attack-motivation',
-    '212': 'attack-resource-level',
-    '213': 'identity-class',
-    '214': 'indicator-label',
-    '215': 'industry-sector',
-    '216': 'malware-label',
-    '218': 'report-label',
-    '219': 'threat-actor-label',
-    '220': 'threat-actor-role',
-    '221': 'threat-actor-sophistication',
-    '222': 'tool-label',
-    '241': 'hash-algo',
-    '242': 'encryption-algo',
-    '243': 'windows-pebinary-type',
-    '244': 'account-type',
-    '270': 'all-external-sources',
-    '271': 'mime-type',
-    '272': 'protocols',
-    '273': 'ipfix',
-    '274': 'http-request-headers',
-    '275': 'socket-options',
-    '276': 'pdf-doc-info',
-    '301': 'network-traffic-ports',
-    '302': 'extref-hashes',
-}
 
 
 def has_cyber_observable_data(instance):
